Restructure files and improve TS
